+++ tests/test_gateway_mqtt.py
@@ -4,8 +4,9 @@ import tempfile
 import time
 from unittest import TestCase, main, mock
 
-from mysensors import ChildSensor, Sensor
 from mysensors.gateway_mqtt import MQTTGateway
+from mysensors.persistence import Persistence
+from mysensors.sensor import ChildSensor, Sensor
 
 
 class TestMQTTGateway(TestCase):
@@ -105,12 +106,15 @@ class TestMQTTGateway(TestCase):
             'ERROR:mysensors.gateway_mqtt:Subscribe to /1/1/1/+/+ failed: '
             'No topic specified, or incorrect topic type.')
 
-    @mock.patch('mysensors.Gateway._save_sensors')
-    @mock.patch('mysensors.Gateway._schedule_save_sensors')
-    def test_start_stop_gateway(self, mock_schedule_save, mock_save):
+    @mock.patch('mysensors.persistence.Persistence.safe_load_sensors')
+    @mock.patch('mysensors.persistence.Persistence.save_sensors')
+    def test_start_stop_gateway(self, mock_save, mock_load):
         """Test start and stop of MQTT gateway."""
-        self.gateway.persistence = True
-        self.gateway.scheduled_save = mock.MagicMock()
+        self.gateway.persistence = Persistence(self.gateway.sensors)
+        mock_cancel_save = mock.MagicMock()
+        mock_schedule_save = mock.MagicMock()
+        mock_schedule_save.return_value = mock_cancel_save
+        self.gateway.persistence.schedule_save_sensors = mock_schedule_save
         self.assertFalse(self.gateway.is_alive())
         sensor = self._add_sensor(1)
         sensor.children[1] = ChildSensor(
@@ -119,14 +123,16 @@ class TestMQTTGateway(TestCase):
         self.gateway.recv('/1/1/2/0/1', '', 0)
         self.gateway.recv('/1/1/1/0/1', '30', 0)
         self.gateway.recv('/1/1/2/0/1', '', 0)
-        self.gateway.start()
-        self.assertTrue(self.gateway.is_alive())
+        self.gateway.start_persistence()
+        assert mock_load.call_count == 1
         assert mock_schedule_save.call_count == 1
+        self.gateway.start()
+        time.sleep(0.05)
+        self.assertTrue(self.gateway.is_alive())
         calls = [
             mock.call('/+/+/0/+/+', self.gateway.recv, 0),
             mock.call('/+/+/3/+/+', self.gateway.recv, 0)]
         self.mock_sub.assert_has_calls(calls)
-        time.sleep(0.05)
         calls = [
             mock.call('/1/1/1/0/1', '20', 0, True),
             mock.call('/1/1/1/0/1', '30', 0, True)]
@@ -134,7 +140,7 @@ class TestMQTTGateway(TestCase):
         self.gateway.stop()
         self.gateway.join(timeout=0.5)
         self.assertFalse(self.gateway.is_alive())
-        assert self.gateway.scheduled_save.cancel.call_count == 1
+        assert mock_cancel_save.call_count == 1
         assert mock_save.call_count == 1
 
     def test_mqtt_load_persistence(self):
@@ -145,12 +151,15 @@ class TestMQTTGateway(TestCase):
         sensor.children[1].values[self.gateway.const.SetReq.V_HUM] = '20'
 
         with tempfile.TemporaryDirectory() as temp_dir:
-            self.gateway.persistence_file = os.path.join(temp_dir, 'file.json')
-            # pylint: disable=protected-access
-            self.gateway._save_sensors()
+            persistence_file = os.path.join(temp_dir, 'file.json')
+            self.gateway.persistence = Persistence(
+                self.gateway.sensors, persistence_file)
+            self.gateway.persistence.save_sensors()
             del self.gateway.sensors[1]
             self.assertNotIn(1, self.gateway.sensors)
-            self.gateway._safe_load_sensors()
+            self.gateway.persistence.safe_load_sensors()
+            # pylint: disable=protected-access
+            self.gateway._init_topics()
         self.assertEqual(
             self.gateway.sensors[1].children[1].id,
             sensor.children[1].id)
